[#280] - Crea StoryService en APIs
